Remove deprecated message dispatcher and request handler files
- Deleted `messageDispatcher.ts` and `requestHandler.ts` as their functionalities are no longer needed.
- Updated the Solana provider to throw a `NotImplementedError` for unimplemented request methods.
- Cleaned up constants and interfaces related to transaction handling for improved code clarity.
