+++ core/inpage-provider/background/interface.ts
@@ -1,5 +1,6 @@
 import { Chain } from '@core/chain/Chain'
 import { ChainOfKind } from '@core/chain/ChainKind'
+import { CoinKey, CoinMetadata, Token } from '@core/chain/coin/Coin'
 import { VaultExport } from '@core/ui/vault/export/core'
 import { Method } from '@lib/utils/types/Method'
 
@@ -23,6 +24,7 @@ export type BackgroundInterface = {
   evmClientRequest: Method<{ method: string; params?: unknown[] }, unknown>
   exportVault: Method<{}, VaultExport>
   getTx: Method<{ chain: Chain; hash: string }, unknown>
+  getTokenMetadata: Method<Token<CoinKey>, CoinMetadata>
 }
 
 export type BackgroundMethod = keyof BackgroundInterface
